# openapi/tokenpilot.openapi.yaml
openapi: 3.1.0
info:
  title: TokenPilot Local Control Plane API
  version: 0.1.0-alpha
  description: >
    Local-first automation surface for TokenPilot. Phase 2 supports ChatGPT direct-drive
    operations for small trusted edits plus asynchronous Codex jobs for complex work. The
    full HTTPS / Custom GPT Actions automation loop remains under validation.
servers:
  - url: https://tokenpilot.example.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    CodexStandaloneCapabilitySnapshot:
      type: object
      additionalProperties: false
      description: Verified standalone App Server capabilities. Probe paths and file contents are omitted.
      properties:
        schemaVersion:
          type: integer
          const: 1
        runtime:
          type: string
          const: codex-app-server
        protocolFamily:
          type: string
          const: app-server-v2
        binarySource:
          oneOf:
            - type: string
            - type: "null"
        binaryVersion:
          oneOf:
            - type: string
            - type: "null"
        serverProtocolVersion:
          oneOf:
            - type: string
            - type: "null"
        probedAt:
          type: string
          format: date-time
        operations:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CodexStandaloneOperationCapability'
        outgoingMethods:
          type: array
          items:
            type: string
        turnStartObserved:
          type: boolean
        directExecutionReady:
          type: boolean
      required:
        - schemaVersion
        - runtime
        - protocolFamily
        - binarySource
        - binaryVersion
        - serverProtocolVersion
        - probedAt
        - operations
        - outgoingMethods
        - turnStartObserved
        - directExecutionReady
    RuntimeCapabilitySnapshot:
      type: object
      additionalProperties: false
      properties:
        available:
          type: boolean
        runtime:
          type: string
          const: codex-app-server
        binarySource:
          oneOf:
            - type: string
            - type: "null"
        binaryVersion:
          oneOf:
            - type: string
            - type: "null"
        protocolFamily:
          type: string
          const: app-server-v2
        serverProtocolVersion:
          oneOf:
            - type: string
            - type: "null"
        stableMethods:
          type: array
          items:
            type: string
        experimentalApiEnabled:
          type: boolean
        standaloneExecution:
          oneOf:
            - $ref: '#/components/schemas/CodexStandaloneCapabilitySnapshot'
            - type: "null"
        unavailableReason:
          type: string
      required:
        - available
        - runtime
        - binarySource
        - binaryVersion
        - protocolFamily
        - serverProtocolVersion
        - stableMethods
        - experimentalApiEnabled
        - standaloneExecution
    RuntimeThreadStatus:
      type: object
      additionalProperties: false
      properties:
        type:
          type: string
        activeFlags:
          type: array
          items:
            type: string
      required: [type]
    RuntimeThreadProjection:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        preview:
          type: string
        modelProvider:
          oneOf:
            - type: string
            - type: "null"
        createdAt:
          oneOf:
            - type: number
            - type: "null"
        updatedAt:
          oneOf:
            - type: number
            - type: "null"
        recencyAt:
          oneOf:
            - type: number
            - type: "null"
        sourceKind:
          oneOf:
            - type: string
            - type: "null"
        status:
          $ref: '#/components/schemas/RuntimeThreadStatus'
        projectId:
          oneOf:
            - type: string
            - type: "null"
        workspaceId:
          oneOf:
            - type: string
            - type: "null"
        repoId:
          oneOf:
            - type: string
            - type: "null"
        parentThreadId:
          oneOf:
            - type: string
            - type: "null"
        agentNickname:
          oneOf:
            - type: string
            - type: "null"
        agentRole:
          oneOf:
            - type: string
            - type: "null"
      required:
        - id
        - preview
        - modelProvider
        - createdAt
        - updatedAt
        - recencyAt
        - sourceKind
        - status
        - projectId
        - workspaceId
        - repoId
        - parentThreadId
        - agentNickname
        - agentRole
    RuntimeCapabilitiesResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        capabilities:
          $ref: '#/components/schemas/RuntimeCapabilitySnapshot'
      required: [ok, capabilities]
    RuntimeThreadListResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeThreadProjection'
        nextCursor:
          oneOf:
            - type: string
            - type: "null"
        backwardsCursor:
          oneOf:
            - type: string
            - type: "null"
      required: [ok, data, nextCursor, backwardsCursor]
    RuntimeThreadReadResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        thread:
          $ref: '#/components/schemas/RuntimeThreadProjection'
      required: [ok, thread]
    RuntimeSessionMutationPayload:
      type: object
      additionalProperties: false
      properties:
        sessionId:
          type: string
          minLength: 1
          maxLength: 160
        threadId:
          type: string
          minLength: 1
          maxLength: 240
        expectedSessionRevision:
          type: integer
          minimum: 1
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [sessionId, threadId, expectedSessionRevision, idempotencyKey]
    RuntimeSessionForkPayload:
      allOf:
        - $ref: '#/components/schemas/RuntimeSessionMutationPayload'
        - type: object
          additionalProperties: false
          properties:
            lastTurnId:
              oneOf:
                - type: string
                  minLength: 1
                  maxLength: 240
                - type: "null"
    RuntimeBindingRecord:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        sessionId:
          type: string
        workspaceId:
          type: string
        runtimeKind:
          type: string
          const: codex-app-server
        externalThreadId:
          type: string
        sourceThreadId:
          oneOf:
            - type: string
            - type: "null"
        relation:
          type: string
          enum: [bound, resumed, forked]
        status:
          type: string
          enum: [active, superseded, released, stale]
        modelProvider:
          oneOf:
            - type: string
            - type: "null"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        revision:
          type: integer
          minimum: 1
      required:
        - id
        - sessionId
        - workspaceId
        - runtimeKind
        - externalThreadId
        - sourceThreadId
        - relation
        - status
        - modelProvider
        - createdAt
        - updatedAt
        - revision
    RuntimeDevelopmentSessionRecord:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        projectId:
          type: string
        workspaceId:
          type: string
        taskId:
          type: string
        title:
          type: string
        mode:
          type: string
          enum: [chat-direct, codex-session, async-agent]
        status:
          type: string
          enum: [idle, running, waiting-approval, handoff-ready, completed, failed]
        activeRuntimeBindingId:
          oneOf:
            - type: string
            - type: "null"
        startedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        endedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        revision:
          type: integer
          minimum: 1
      required:
        - id
        - projectId
        - workspaceId
        - taskId
        - title
        - mode
        - status
        - activeRuntimeBindingId
        - startedAt
        - updatedAt
        - endedAt
        - revision
    RuntimeBindingMutationResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        binding:
          $ref: '#/components/schemas/RuntimeBindingRecord'
        session:
          $ref: '#/components/schemas/RuntimeDevelopmentSessionRecord'
        thread:
          $ref: '#/components/schemas/RuntimeThreadProjection'
        replayed:
          type: boolean
      required: [ok, binding, session, thread, replayed]
    RuntimeTurnStartPayload:
      type: object
      additionalProperties: false
      properties:
        sessionId:
          type: string
          minLength: 1
          maxLength: 160
        text:
          type: string
          minLength: 1
          maxLength: 50000
        expectedSessionRevision:
          type: integer
          minimum: 1
        expectedTaskRevision:
          type: integer
          minimum: 1
        leaseDurationSeconds:
          type: integer
          minimum: 60
          maximum: 3600
          default: 900
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required:
        - sessionId
        - text
        - expectedSessionRevision
        - expectedTaskRevision
        - idempotencyKey
    RuntimeTurnInterruptPayload:
      type: object
      additionalProperties: false
      properties:
        runId:
          type: string
          minLength: 1
          maxLength: 160
        expectedRunRevision:
          type: integer
          minimum: 1
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [runId, expectedRunRevision, idempotencyKey]
    RuntimeApprovalRespondPayload:
      type: object
      additionalProperties: false
      properties:
        approvalId:
          type: string
          minLength: 1
          maxLength: 160
        expectedRevision:
          type: integer
          minimum: 1
        decision:
          type: string
          enum: [accept, decline, cancel]
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 128
          pattern: '^[A-Za-z0-9._:-]+$'
      required: [approvalId, expectedRevision, decision, idempotencyKey]
    RuntimeTurnProjection:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        status:
          type: string
        startedAt:
          oneOf:
            - type: number
            - type: "null"
        completedAt:
          oneOf:
            - type: number
            - type: "null"
        durationMs:
          oneOf:
            - type: number
            - type: "null"
        errorCode:
          oneOf:
            - type: string
            - type: "null"
      required: [id, status, startedAt, completedAt, durationMs, errorCode]
    RuntimeRunRecord:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
        sessionId:
          type: string
        workspaceId:
          type: string
        runtimeBindingId:
          type: string
        threadId:
          type: string
        externalTurnId:
          oneOf:
            - type: string
            - type: "null"
        status:
          type: string
          enum: [starting, running, waiting-approval, completed, failed, interrupted, stale]
        inputHash:
          type: string
          description: SHA-256 hash of the submitted input. The raw input is not exposed here.
        inputLength:
          type: integer
          minimum: 0
        handoffId:
          type: string
        evidenceBundleId:
          type: string
        writerLeaseId:
          type: string
        modelLoopOwner:
          type: string
          const: codex
        approvalPolicy:
          type: string
          const: on-request
        startedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        errorCode:
          oneOf:
            - type: string
            - type: "null"
        revision:
          type: integer
          minimum: 1
      required:
        - id
        - sessionId
        - workspaceId
        - runtimeBindingId
        - threadId
        - externalTurnId
        - status
        - inputHash
        - inputLength
        - handoffId
        - evidenceBundleId
        - writerLeaseId
        - modelLoopOwner
        - approvalPolicy
        - startedAt
        - updatedAt
        - completedAt
        - errorCode
        - revision
    RuntimeApprovalRecord:
      type: object
      additionalProperties: false
      description: Public approval projection. Internal App Server request handles and raw private request parameters are omitted.
      properties:
        id:
          type: string
        runId:
          type: string
        sessionId:
          type: string
        workspaceId:
          type: string
        threadId:
          type: string
        turnId:
          type: string
        itemId:
          oneOf:
            - type: string
            - type: "null"
        requestMethod:
          type: string
        kind:
          type: string
          enum: [command-execution, file-change, permissions, unsupported]
        status:
          type: string
          enum: [pending, responded, resolved, cancelled, stale]
        publicSummary:
          type: object
          additionalProperties: true
        decision:
          oneOf:
            - type: object
              additionalProperties: true
            - type: "null"
        receivedAt:
          type: string
          format: date-time
        respondedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        resolvedAt:
          oneOf:
            - type: string
              format: date-time
            - type: "null"
        revision:
          type: integer
          minimum: 1
      required:
        - id
        - runId
        - sessionId
        - workspaceId
        - threadId
        - turnId
        - itemId
        - requestMethod
        - kind
        - status
        - publicSummary
        - decision
        - receivedAt
        - respondedAt
        - resolvedAt
        - revision
    RuntimeEventRecord:
      type: object
      additionalProperties: false
      description: Append-only public-safe runtime event. Raw prompts, cwd, patches, command output, and private approval payloads are omitted.
      properties:
        sequence:
          type: integer
          minimum: 1
        id:
          type: string
        runId:
          oneOf:
            - type: string
            - type: "null"
        sessionId:
          type: string
        workspaceId:
          type: string
        threadId:
          type: string
        turnId:
          oneOf:
            - type: string
            - type: "null"
        itemId:
          oneOf:
            - type: string
            - type: "null"
        method:
          type: string
        category:
          type: string
          enum: [lifecycle, approval, item, warning, error, other]
        publicPayload:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required:
        - sequence
        - id
        - runId
        - sessionId
        - workspaceId
        - threadId
        - turnId
        - itemId
        - method
        - category
        - publicPayload
        - createdAt
    RuntimeTurnStartResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        run:
          $ref: '#/components/schemas/RuntimeRunRecord'
        lease:
          type: object
          additionalProperties: true
        handoff:
          type: object
          additionalProperties: true
        evidenceBundle:
          type: object
          additionalProperties: true
        session:
          $ref: '#/components/schemas/RuntimeDevelopmentSessionRecord'
        task:
          type: object
          additionalProperties: true
        turn:
          $ref: '#/components/schemas/RuntimeTurnProjection'
        replayed:
          type: boolean
      required: [ok, run, lease, handoff, evidenceBundle, session, task, turn, replayed]
    RuntimeTurnInterruptResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        run:
          $ref: '#/components/schemas/RuntimeRunRecord'
        session:
          $ref: '#/components/schemas/RuntimeDevelopmentSessionRecord'
        lease:
          type: object
          additionalProperties: true
        replayed:
          type: boolean
      required: [ok, run, session, lease, replayed]
    RuntimeApprovalRespondResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        approval:
          $ref: '#/components/schemas/RuntimeApprovalRecord'
        run:
          $ref: '#/components/schemas/RuntimeRunRecord'
        session:
          $ref: '#/components/schemas/RuntimeDevelopmentSessionRecord'
        replayed:
          type: boolean
      required: [ok, approval, run, session, replayed]
    RuntimeEventsResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        events:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeEventRecord'
        nextSequence:
          oneOf:
            - type: integer
              minimum: 1
            - type: "null"
      required: [ok, events, nextSequence]
paths:
...
  /api/runtime/codex/capabilities:
    get:
      operationId: getCodexRuntimeCapabilities
      summary: Read Codex App Server capabilities
      description: Returns a public-safe availability and protocol snapshot. Binary paths and local diagnostics are not exposed.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Codex runtime capability snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeCapabilitiesResponse'
  /api/runtime/codex/threads:
    get:
      operationId: listCodexRuntimeThreads
      summary: List Codex thread metadata
      description: Lists public-safe Codex thread metadata. Raw cwd, instruction sources, rollout paths, and turn contents are omitted.
      security:
        - bearerAuth: []
      parameters:
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: workspaceId
          in: query
          required: false
          schema:
            type: string
          description: TokenPilot workspace id. The private workspace path is resolved only inside the local runtime.
        - name: searchTerm
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 500
        - name: archived
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Public-safe thread list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeThreadListResponse'
  /api/runtime/codex/threads/{threadId}:
    get:
      operationId: getCodexRuntimeThread
      summary: Read Codex thread metadata
      description: Reads one public-safe thread metadata projection. Turn history remains unavailable until TokenPilot defines and verifies a safe projection.
      security:
        - bearerAuth: []
      parameters:
        - name: threadId
          in: path
          required: true
          schema:
            type: string
        - name: includeTurns
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Must remain false in the current read-only adapter. True returns CAPABILITY_UNAVAILABLE.
      responses:
        "200":
          description: Public-safe thread metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeThreadReadResponse'
  /api/runtime/codex/sessions/bind:
    post:
      operationId: bindCodexRuntimeSession
      summary: Bind an existing Codex thread
      description: Validates TokenPilot session revision and workspace identity, then binds an existing Codex thread without starting a Codex turn. Repeated requests with the same idempotency key replay the original result.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeSessionMutationPayload'
      responses:
        "200":
          description: Created or replayed runtime binding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeBindingMutationResponse'
  /api/runtime/codex/sessions/resume:
    post:
      operationId: resumeCodexRuntimeSession
      summary: Resume and bind a Codex thread
      description: Resumes an existing Codex App Server thread and persists the binding with two-phase idempotency. This operation does not start a Codex turn. An uncertain external result remains pending to prevent duplicate resume side effects.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeSessionMutationPayload'
      responses:
        "200":
          description: Resumed or replayed runtime binding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeBindingMutationResponse'
  /api/runtime/codex/sessions/fork:
    post:
      operationId: forkCodexRuntimeSession
      summary: Fork and bind a Codex thread
      description: Creates a durable non-ephemeral Codex thread fork and binds it to the TokenPilot session with two-phase idempotency. An optional lastTurnId limits the fork boundary. This operation does not start a Codex turn.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeSessionForkPayload'
      responses:
        "200":
          description: Forked or replayed runtime binding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeBindingMutationResponse'
  /api/runtime/codex/turns/start:
    post:
      operationId: startCodexRuntimeTurn
      summary: Start an explicit Codex turn
      description: Starts one Codex model loop after validating Runtime Binding, revisions, Writer Lease, Git checkpoint, Evidence, and Handoff. Approval is fixed to on-request user review. cwd, model, sandbox, approval, and instruction overrides are unavailable. Retries use two-phase idempotency.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeTurnStartPayload'
      responses:
        "200":
          description: Started or replayed Codex turn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeTurnStartResponse'
  /api/runtime/codex/turns/interrupt:
    post:
      operationId: interruptCodexRuntimeTurn
      summary: Interrupt an active Codex turn
      description: Explicitly interrupts one active bound Codex turn, releases its Writer Lease, moves the Development Session to handoff-ready, and records an append-only lifecycle event. The operation is idempotent across REST and MCP retries.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeTurnInterruptPayload'
      responses:
        "200":
          description: Interrupted or replayed Codex turn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeTurnInterruptResponse'
  /api/runtime/codex/approvals/respond:
    post:
      operationId: respondToCodexRuntimeApproval
      summary: Respond to a pending Codex approval
      description: Sends an explicit accept, decline, or cancel decision for one pending command-execution or file-change approval. Session-wide approval and permission escalation are intentionally unavailable. Internal App Server request handles and raw private request parameters are never returned by this API.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeApprovalRespondPayload'
      responses:
        "200":
          description: Responded or replayed approval decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeApprovalRespondResponse'
  /api/runtime/codex/events:
    get:
      operationId: readCodexRuntimeEvents
      summary: Read public-safe Codex runtime events
      description: Reads append-only lifecycle, approval, item, warning, and error events for one TokenPilot Session or Runtime Run. Raw prompts, command output, cwd, file patches, private approval payloads, and App Server request handles are omitted.
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: query
          required: false
          schema:
            type: string
          description: Filter by TokenPilot Development Session. sessionId or runId is required.
        - name: runId
          in: query
          required: false
          schema:
            type: string
          description: Filter by TokenPilot Runtime Run. sessionId or runId is required.
        - name: afterSequence
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Return events after this append-only sequence cursor.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Public-safe runtime event page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeEventsResponse'